refactor(game-color):LetterColorPlay게임 컴포넌트 폴더로 이동
