+++ src/pages/Games/LetterColor/components/letterColorPlay.module.css
